% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\name{C.partitioning}
\alias{C.partitioning}
\title{Simulating how the carbon is partitioned}
\usage{
C.partitioning(AG.simulation, parameters)
}
\arguments{
\item{AG.simulation}{output from from \code{\link{Sim.aboveground}}}

\item{parameters}{The parameter matrix from \code{\link{parameters}}}
}
\value{
a data frame with #' \itemize{
  \item C in needles
  \item C in branches
  \item C in stem V
  \item C in stem B
  \item C in fine roots
  \item C in roots (>5 cm)
  \item C in roots (<5 cm)
  \item C in stumps
}
for each year of the simulation
}
\description{
The C.partitioning function describes how the carbon is partitioned in the stand, based on the model
parameter matrix from \code{\link{parameters}} and on the results from \code{\link{Sim.aboveground}}
}
