% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\name{Harvest.residuals}
\alias{Harvest.residuals}
\title{Simulating the harvest residuals from the C partitions}
\usage{
Harvest.residuals(C.partitions, management.plan, parameters)
}
\arguments{
\item{C.partitions}{output from from \code{\link{C.partitioning}}}

\item{management.plan}{output from from \code{\link{management.plan}}}

\item{parameters}{The parameter matrix from \code{\link{parameters}}}
}
\value{
a data frame with #' \itemize{
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
  \item C !!! to be defined better
}
for each year of the simulation
}
\description{
The Harvest.residuals function describes the litter production based on the model
parameter matrix from \code{\link{parameters}} and on the results from \code{\link{C.partitioning}}
}
