% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\name{Litter.production}
\alias{Litter.production}
\title{Simulating the litter production from C partitions}
\usage{
Litter.production(C.partitions, Lat, parameters)
}
\arguments{
\item{C.partitions}{output from from \code{\link{C.partitioning}}}

\item{Lat}{The latitude of the simulated site.}

\item{parameters}{The parameter matrix from \code{\link{parameters}}}
}
\value{
a data frame with #' \itemize{
  \item C litter from fine roots
  \item C litter from branches
}
for each year of the simulation
}
\description{
The Litter.production function describes the litter production based on the model
parameter matrix from \code{\link{parameters}} and on the results from \code{\link{C.partitioning}}
}
