% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\name{management.plan}
\alias{management.plan}
\title{The function for determining the appropriate management plan.}
\usage{
management.plan(SI, Lat)
}
\arguments{
\item{SI}{The site index.}

\item{Lat}{The latitude of the simulated site.}
}
\value{
a data frame with site selection ID for the other functions of the package, diameter at breast height T., and ages of the four thinnings
}
\description{
The management.plan functions contains a series of vectors with management data from Sweden.
It selects the most appropriate values based on site index and latitude, and determines the
values for the management plan.
}
