% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Just a data frame with the model parameters.}
\format{
An object of class \code{data.frame} with 1 rows and 13 columns.
}
\usage{
parameters
}
\description{
Just a data frame with the model parameters.
}
\keyword{datasets}
