% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.0.1.R
\name{Remaining.OM}
\alias{Remaining.OM}
\title{The core function for simulating the decay of OM}
\usage{
Remaining.OM(Lat, parameters, tmax, timestep)
}
\arguments{
\item{Lat}{The latitude of the simulated site.}

\item{parameters}{The parameter matrix from \code{\link{parameters}}}

\item{tmax}{last harvest (??? to be checked...)}

\item{timestep}{time steps for the simulation}
}
\value{
"G", the fraction of OM remaining (??? to be checked...)
}
\description{
The Remaining.OM function simulates the organic matter decay. It is the core function for organic matter kinetic,
to be used in the simulations in another function
}
